refactor: move llm session usage helper
