+++ apps/backend/src/agent-core/llm-session/helpers.ts
@@ -1,0 +1,10 @@
+import type {LlmSessionUsage} from './types.js';
+
+export function createEmptyLlmSessionUsage(): LlmSessionUsage {
+  return {
+    currentContextInputTokens: 0,
+    sessionInputTokens: 0,
+    sessionOutputTokens: 0,
+    sessionCacheReadInputTokens: 0,
+  };
+}
